Add files via upload
Visual DRA in Collaboration (Dynamic Role Assignment)
Please cite:
[1]	H. Zhu and Shawn Mason, “A Visualization Tool for Adaptive Collaboration”, The IEEE Int’l Conf. on Systems, Man and Cybernetics (SMC’16), Budapest, Hungary, Oct. 9-12, 2016, pp. 1005-1010.
[2]	H. Zhu, “Adaptive Collaboration while Agents’ Performance Changes Periodically”, in Proc. of ACM/IEEE Int’l Conf. on Collaborative Technologies and Systems, Atlanta, GA, USA, June 6-9, 2015, pp. 106-112.
[3]	H. Zhu, “Adaptive Collaboration Systems”, IEEE Systems, Man, and Cybernetics Magazine, vol. 1, no. 4, Oct. 2015, pp. 8-15.
[4]	H. Zhu, M. Hou, and M.C. Zhou, “Adaptive Collaboration Based on the E-CARGO Model”, Int’l J. of Agent Technologies and Systems, vol. 4, no.1, 2012, pp. 59-76.
[5]	H. Zhu, M.C. Zhou, and R. Alkins, “Group Role Assignment via a Kuhn-Munkres Algorithm-based Solution”, IEEE Trans. on Systems, Man, and Cybernetics, Part A: Systems and Humans, vol. 42, no. 3, May 2012, pp. 739-750. 
[6]	H. Zhu, and M. Zhou, “Role-Based Collaboration and its Kernel Mechanisms,” IEEE Trans. on Systems, Man, and Cybernetics, Part C: Applications and Reviews, vol. 36, no. 4, 2006, pp. 578-589. 

Version: 1.0.1
Build: beta 1.0001
Build Date: Jan 04, 2016
Developed By: Shawn Mason
Developed For: Dr. Haibin Zhu

This application is graphical user interface (GUI) that allows the visualization of Dynamic Role Assignment (DRA) in Collaboration. Through the use of the E-CARGO model, a user is able to visualize the performance effects of agent to role assignments over a period of time.

[License]
Application license:
The program will falls under the GPL v3 license and be held accountable to the terms of the GPL v3 license.

Images license:
The images utilized in this application have been sourced online. The use of this software for educational and/or internal use allows the owner of this application to utilize these images without copyright infringement. Should the application be distributed, then the images must be replaced with legally obtained replacement images and/or originally created images.

Algorithm license:
The algorithm was originally designed by Gary Baker in 2007 and is licensed under GPL v3. In addition, the algorithm was modified by Dr. Haibin Zhu in 2009 and by Shawn Mason in 2015 and falls under GPL v3.